Forms reorganized. Games app related forms moved to games.forms.py. 'get_game_year_choices' get updates and docs. 'SearchFormMixin' renamed to 'GameFiltersMixin'. Removed 'SearchBarForm' inheritance from 'GameFiltersMixin'.
